executor/pyproject.toml
[project]
name = "executor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docker>=7.1.0",
    "influxdb-client>=1.46.0",
    "psycopg2-binary>=2.9.9",
]
